fix linting and bug
